Module covid19_vaccine_model::vaccination_model

The module implementing everything from the model and the solver.

How it is organized:

All the model parameters and equations are included in the submodule model. The solver submodule includes the Runge-Kutta 4 solver and the PD control system and can write the numerical solutions to a folder. The submodule age_group_state_vector implements some data structures to store the system state.

How to use it:

create the model with the global parameters and add the individual age groups to it

initialize the model and prepare vaccination rates

create the solver

initialize the solver

run the solver using a specified scenario

write the data to a file

Structs

AgeGroup
A collection of all the parameters for a given age group

AgeGroupStateVector
Used to store all the information on the state of one age group

Model
A collection of global parameters (especially the vaccination parameters) and all the age groups. Includes the dif. eqs.

Solver
Runge-Kutta 4 solver and PD control system for the model.