Struct covid19_vaccine_model::vaccination_model::solver::Solver

pub struct Solver {
    pub model: Model,
    pub dt: f64,
    pub t0: f64,
    pub initials: Vec<AgeGroupStateVector>,
    pub Rt_initial: f64,
    pub time: Vec<f64>,
    pub H: Vec<f64>,
    pub Rt: Vec<f64>,
    pub states: Vec<Vec<AgeGroupStateVector>>,
    pub N: Vec<f64>,
    pub N_obs: Vec<f64>,
    pub index: usize,
}

Runge-Kutta 4 solver and PD control system for the model.

Fields
model: Model

Model parameters

dt: f64

Step size for the RK4 solver

t0: f64

Initial time (t0=0 indicates the start of the vaccination programe, i.e. end of December 2020)

initials: Vec<AgeGroupStateVector>

Initial values for the system state (all compartments)

Rt_initial: f64

Initial $R_t$ value

time: Vec<f64>

Result vector for the times $t$

H: Vec<f64>

Result vector for the H helper variable

H is a helper variable and is defined as $\int_{t-\tau}^t \frac{R_t'}{M}\sum_{i,\nu}\bar\gamma_i I^\nu_i(t')dt'$. From its integral defintion it can be easily converted into a differential form. The integral can thus always be calculated with the dif. eq. solver alongside all other variables. In the end this variable can then be easily used to calculate the $p_i(t)$ from the manuscript, the fraction of vaccinated people that get infected before developing their immunisation. Essentially it was introduced to convert the system of delay differential equations depending on the delayed time continuum from t-tau to t into one depending on only the system state at the discrete delayed time t-tau by adding one more differential equation.

Rt: Vec<f64>

Result vector for the raw $R_t$ used in the dif. eqs. (not test-trace-and-isolate (TTI) corrected)

states: Vec<Vec<AgeGroupStateVector>>

Result vector for the system states (compartments)

N: Vec<f64>

Result vector for the total daily infections (not age resolved, not delayed)

N_obs: Vec<f64>

Result vector for the total daily infections (not age resolved, delayed by observation kernel K=[0.0,0.0,0.5,0.3,0.1,0.1])

index: usize

Current index (where we are in the result vectors)

Implementations

impl Solver

pub fn initialize(&mut self)

Initializes the solver. Clears all result arrays and writes initial values into them, sets index to 0.

pub fn controlled_run(
&mut self,
T: f64,
change_points: &[(f64, f64, f64, f64, usize, f64)]
) -> f64

Runs the simulation for a timespan T. Recieves a series of control problems seperated by change points.

How to use

solver.controlled_run(T, &[(t0, min0, max0, max_slope0, control0, aim0), (t1, min1, max1, max_slope1, control1, aim1), ...]);

runs the solver for a timespan T. In the beginning the control approach control0 is chosen to aim at set point aim0. With a minimal ((test-trace-and-isolate (TTI) corrected) $R_t$ min0 and a maximal value max0. $R_t$ is allowed to increase only with a maximal slope max_slope0. The system is run until time t0 is reached and the parameters are updated to min1, max1, max_slope1, control1, aim1.

control=0 means we aim at stable daily infections given by the set target value aim.

control=1 means we aim at stable ICU occupancy given by the set target value aim.

control=2 is the same as control=1 but it integrates the time where ICU is at the capacity limit (i.e. 70% close to aim) and returns that time.

pub fn run_rk4(
&self,
T: f64,
time: &mut Vec<f64>,
H: &mut Vec<f64>,
Rt: &mut Vec<f64>,
states: &mut Vec<Vec<AgeGroupStateVector>>,
N: &mut Vec<f64>,
time_history: &Vec<f64>,
H_history: &Vec<f64>,
Rt_history: &Vec<f64>,
states_history: &Vec<Vec<AgeGroupStateVector>>,
R: f64
) -> usize

Solves the system of delay diff. eqs. for a timespan T using Runge-Kutta 4. Saves the results in time, H, Rt, states and N. Uses the respective history arrays if the delays reach out of the current simulation.

Returns the index in the result arrays in the end for easy access.

pub fn write_to_disk(&self, foldername: &str, write_every: usize) -> Result<()>

Writes the results to a folder "./data/foldername/". To reduce file size it only writes every write_every value of the results.

Panics if write or file creation failed somewhere, i.e. if the directory does not exist.