Module covid19_vaccine_model::vaccination_model

The module implementing everything from the model and the solver.

How it is organized:

All the model parameters and equations are included in the submodule model. The solver submodule includes the Runge-Kutta 4 solver and the PD control system and can write the numerical solutions to a folder.

How to use it:

create the model with the global parameters and add the individual age groups to it

initialize the model and prepare vaccination rates

create the solver

initialize the solver

run the solver using a specified scenario

write the data to a file

Re-exports

pub use model::Model;

pub use model::AgeGroup;

pub use solver::Solver;

pub use age_group_state_vector::AgeGroupStateVector;

Modules

age_group_state_vector
Everything needed to store the state of the system.

model
Everything related to the model. Including collections of global and age-group specific parameters and the differential equations.

solver
Runge-Kutta 4 solver and PD control system for the model.